+++ config/releases.exs
@@ -40,7 +40,8 @@ host =
 config :oli,
   email_from_name: System.get_env("EMAIL_FROM_NAME", "OLI Torus"),
   email_from_address: System.get_env("EMAIL_FROM_ADDRESS", "admin@example.edu"),
-  email_reply_to: System.get_env("EMAIL_REPLY_TO", "admin@example.edu")
+  email_reply_to: System.get_env("EMAIL_REPLY_TO", "admin@example.edu"),
+  slack_webhook_url: System.get_env("SLACK_WEBHOOK_URL")
 
 # Configure reCAPTCHA
 config :oli, :recaptcha,