+++ lib/oli/email.ex
@@ -25,7 +25,24 @@ defmodule Oli.Email do
     base_email()
     |> from(name <> "<" <> from_email <> ">")
     |> put_header("Reply-To", name <> " <" <> from_email <> ">")
+    |> put_layout({OliWeb.LayoutView, :help_email})
     |> to(help_desk_email)
+    |> subject(subject)
+    |> render(view, assigns)
+    |> html_text_body()
+  end
+
+  @doc """
+  Creates a generic email with an html and text body.
+  Returns a Bamboo.Email struct ready for delivery.
+  ## Examples:
+    iex> create_email("someone@example.com", "Example Email", "template.html", assigns)
+    %Bamboo.Email{}
+  """
+  @spec create_email(String.t(), String.t(), String.t(), map()) :: Bamboo.Email.t()
+  def create_email(recipient_email, subject, view, assigns) do
+    base_email()
+    |> to(recipient_email)
     |> subject(subject)
     |> render(view, assigns)
     |> html_text_body()
@@ -39,8 +56,8 @@ defmodule Oli.Email do
     new_email()
     |> from({from_email_name, from_email_address})
     |> put_header("Reply-To", email_reply_to)
-    # render both email.html and email.text layouts using :help_email
-    |> put_layout({OliWeb.LayoutView, :help_email})
+    # render both email.html and email.text layouts using :email
+    |> put_layout({OliWeb.LayoutView, :email})
   end
 
   def html_text_body(email) do