+++ lib/oli/institutions/institution.ex
@@ -9,8 +9,6 @@ defmodule Oli.Institutions.Institution do
     field :name, :string
     field :timezone, :string
 
-    belongs_to :author, Oli.Accounts.Author, foreign_key: :author_id
-
     # LTI 1.3 Deployments
     has_many :registrations, Oli.Lti_1p3.Registration
     has_many :sections, Oli.Delivery.Sections.Section
@@ -28,7 +26,6 @@ defmodule Oli.Institutions.Institution do
       :institution_email,
       :institution_url,
       :timezone,
-      :author_id,
     ])
     |> validate_required([
       :name,
@@ -36,7 +33,6 @@ defmodule Oli.Institutions.Institution do
       :institution_email,
       :institution_url,
       :timezone,
-      :author_id,
     ])
   end
 end