+++ lib/oli_web/router.ex
@@ -298,6 +298,8 @@ defmodule OliWeb.Router do
     post "/test", LtiController, :test
 
     get "/developer_key.json", LtiController, :developer_key_json
+
+    post "/register", LtiController, :request_registration
   end
 
   scope "/course", OliWeb do
@@ -312,7 +314,6 @@ defmodule OliWeb.Router do
 
     post "/section", DeliveryController, :create_section
     get "/signout", DeliveryController, :signout
-    get "/open_and_free", DeliveryController, :list_open_and_free
 
     get "/unauthorized", DeliveryController, :unauthorized
 
@@ -352,6 +353,8 @@ defmodule OliWeb.Router do
     get "/invite", InviteController, :index
     post "/invite", InviteController, :create
 
+    put "/approve_registration", InstitutionController, :approve_registration
+    delete "/pending_registration/:id", InstitutionController, :remove_registration
   end
 
   scope "/project", OliWeb do