+++ test/oli_web/controllers/institution_controller_test.exs
@@ -1,10 +1,14 @@
 defmodule OliWeb.InstitutionControllerTest do
   use OliWeb.ConnCase
 
+  import ExUnit.CaptureLog
+
   alias Oli.Repo
   alias Oli.Accounts
   alias Oli.Accounts.Author
   alias Oli.Institutions
+  alias Oli.Lti_1p3.{Registration, Deployment}
+  alias Oli.Institutions.PendingRegistration
 
   @create_attrs %{country_code: "some country_code", institution_email: "some institution_email", institution_url: "some institution_url", name: "some name", timezone: "some timezone"}
   @update_attrs %{country_code: "some updated country_code", institution_email: "some updated institution_email", institution_url: "some updated institution_url", name: "some updated name", timezone: "some updated timezone"}
@@ -40,6 +44,25 @@ defmodule OliWeb.InstitutionControllerTest do
     test "renders errors when data is invalid", %{conn: conn} do
       conn = post(conn, Routes.institution_path(conn, :create), institution: @invalid_attrs)
       assert html_response(conn, 200) =~ "Register Institution"
+    end
+  end
+
+  describe "show institution" do
+    setup [:create_institution]
+
+    test "renders institution details", %{conn: conn, institution: institution} do
+      conn = get(conn, Routes.institution_path(conn, :show, institution))
+      assert html_response(conn, 200) =~ "some name"
+    end
+
+    test "renders institution registration details", %{conn: conn, institution: institution} do
+      jwk = jwk_fixture()
+      %Registration{id: registration_id} = registration_fixture(%{institution_id: institution.id, tool_jwk_id: jwk.id})
+      %Deployment{deployment_id: deployment_id} = deployment_fixture(%{registration_id: registration_id})
+
+      conn = get(conn, Routes.institution_path(conn, :show, institution))
+      assert html_response(conn, 200) =~ "some issuer - some client_id"
+      assert html_response(conn, 200) =~ deployment_id
     end
   end
 
@@ -84,6 +107,25 @@ defmodule OliWeb.InstitutionControllerTest do
 
         get(conn, Routes.institution_path(conn, :show, institution))
       end
+    end
+  end
+
+  describe "approve registration" do
+    setup [:create_institution]
+
+    test "approves the chosen registration", %{conn: conn} do
+      pending_registration = pending_registration_fixture()
+      pending_registration_attrs = %{}
+        |> Map.merge(PendingRegistration.institution_attrs(pending_registration))
+        |> Map.merge(PendingRegistration.registration_attrs(pending_registration))
+
+
+      assert capture_log(fn ->
+        conn = put(conn, Routes.institution_path(conn, :approve_registration), %{"pending_registration" => pending_registration_attrs})
+        assert redirected_to(conn) == Routes.institution_path(conn, :index) <> "#pending-registrations"
+
+        assert Institutions.count_pending_registrations() == 0
+      end) =~ "This message cannot be sent because SLACK_WEBHOOK_URL is not configured"
     end
   end